ml/requirements.txt
tensorflow>=2.10.0
numpy>=1.23.0
Pillow>=9.0.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
